+++ app/src/main/java/moe/antimony/hoshi/importing/ImportFileType.kt
@@ -69,7 +69,7 @@ data class ImportFileType(
         )
 
         val TtuBookDataBackup = ImportFileType(
-            description = "TTU bookdata backup",
+            description = "ッツ bookdata backup",
             extensions = listOf("zip"),
             mimeTypes = arrayOf("application/zip", "application/x-zip-compressed", "application/octet-stream"),
             unsupportedMessageRes = R.string.import_select_ttu_bookdata_backup,